chore(logger): revert moving to own file
