--- src/logger.js
@@ -1,3 +1,0 @@
-import {getLogger} from 'aurelia-logging';
-
-export const logger = getLogger('aurelia-orm');
